[Back] fix: 변수 type 수정

+++ BE/src/main/java/com/kkirikkiri/domain/book/dto/LibraryResponse.java
@@ -11,7 +11,7 @@ public class LibraryResponse {
     private String author;
     private String summary;
     private String imageURL;
-    private Long download;
+    private Integer download;
     private String date;
     private Boolean possession;
